# docs/openapi_fe.yaml
openapi: 3.0.3
info:
  title: RinnoGen API — Frontend-Facing
  description: |
    Public-facing endpoints consumed by the RinnoGen frontend.
    Admin, webhook, and internal service routes are excluded.

    **Response Envelope**:
    - Success: `{status, code, message, data}`
    - Error:   `{status, code, message, error, details}`

    **Auth**: Protected routes require `Authorization: Bearer <JWT>`.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiSuccess:
      type: object
      required: [status, code, message, data]
      properties:
        status: { type: string, enum: [success] }
        code: { type: integer }
        message: { type: string }
        data: { type: object }
      example:
        status: success
        code: 200
        message: "Endpoint specific success message"
        data: {}
    ApiError:
      type: object
      required: [status, code, message, error]
      properties:
        status: { type: string, enum: [error] }
        code: { type: integer }
        message: { type: string }
        error: { type: string }
        details: { type: object, nullable: true }
      example:
        status: error
        code: 400
        message: Invalid request body
        error: BAD_REQUEST
        details: null

    # --- Auth ---
...
    ProblemListItem:
      type: object
      properties:
        id: { type: integer, example: 1 }
        slug: { type: string, example: sum-two-numbers }
        title: { type: string, example: "Sum Two Numbers" }
        difficulty: { type: string, enum: [Easy, Medium, Hard], example: Easy }
        orderIndex: { type: integer, example: 1 }
        acceptanceRate: { type: number, example: 75.5 }
    ProblemListData:
      type: object
      properties:
        problems:
          type: array
          items: { $ref: "#/components/schemas/ProblemListItem" }

    # --- Problems ---
    ProblemDetail:
      type: object
      properties:
        executionMode:
          type: string
          description: The execution mode (e.g. stdin_stdout or function)
        functionName:
          type: string
          nullable: true
          description: Name of the function to be implemented
        initialCode:
          type: string
          nullable: true
          description: Starting code template for the user
        id: { type: integer, example: 1 }
        slug: { type: string, example: sum-two-numbers }
        title: { type: string, example: "Sum Two Numbers" }
        difficulty: { type: string, example: Easy }
        problemMd:
          {
            type: string,
            example: "## Sum Two Numbers\n\nGiven two integers, return their sum.",
          }
        timeLimitMs: { type: integer, example: 1000 }
        memoryLimitMb: { type: integer, example: 128 }
        acceptanceRate: { type: number, example: 75.5 }
        sampleTestCases:
          type: array
          items:
            type: object
            properties:
              input: { type: string, example: "2 3" }
              output: { type: string, example: "5" }
    ProblemDetailData:
      type: object
      properties:
        problem: { $ref: "#/components/schemas/ProblemDetail" }
...
  responses:
    BadRequest:
      description: Invalid request (malformed body or missing required fields)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
          example:
            status: error
            code: 400
            message: Invalid request body
            error: BAD_REQUEST
            details: null
...
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
          example:
            status: error
            code: 404
            message: Not found
            error: NOT_FOUND
            details: null
...
paths:
...
  /lessons/{id}/problems:
    get:
      tags: [Curriculum]
      security: []
      summary: List problems for a lesson
      description: Returns published problems associated with a lesson, ordered by the lesson-specific orderIndex.
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: integer },
            example: 1,
          }
      responses:
        "200":
          description: List of problems
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiSuccess" }
                  - properties:
                      data: { $ref: "#/components/schemas/ProblemListData" }
              example:
                status: success
                code: 200
                message: "Endpoint specific success message"
                data:
                  problems:
                    - id: 1
                      slug: sum-two-numbers
                      title: "Sum Two Numbers"
                      difficulty: Easy
                      orderIndex: 1
                      acceptanceRate: 75.5
                    - id: 2
                      slug: hello-world
                      title: "Hello World"
                      difficulty: Easy
                      orderIndex: 2
                      acceptanceRate: 90.0
        "400": { $ref: "#/components/responses/BadRequest" }

  # ──────────────────────────────────────────────
...
  /problems/{slug}:
    get:
      tags: [Problems]
      security: []
      summary: Get problem by slug
      description: Returns full problem details including Markdown description, time/memory limits, acceptance rate, and sample test cases.
      parameters:
        - {
            name: slug,
            in: path,
            required: true,
            schema: { type: string },
            example: sum-two-numbers,
          }
      responses:
        "200":
          description: Problem details
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiSuccess" }
                  - properties:
                      data: { $ref: "#/components/schemas/ProblemDetailData" }
              example:
                status: success
                code: 200
                message: "Endpoint specific success message"
                data:
                  problem:
                    id: 1
                    slug: sum-two-numbers
                    title: "Sum Two Numbers"
                    difficulty: Easy
                    problemMd: "## Sum Two Numbers\n\nGiven two integers, return their sum."
                    timeLimitMs: 1000
                    memoryLimitMb: 128
                    acceptanceRate: 75.5
                    sampleTestCases:
                      - input: "2 3"
                        output: "5"
                      - input: "-1 1"
                        output: "0"
        "404": { $ref: "#/components/responses/NotFound" }
...